// src/enums/user.enum.ts
export enum UserStatus {
  ACTIVE = "active",
  WITHDRAWN = "withdrawn",
}

export enum UserType {
  ADMIN = "admin",
  MEMBER = "member",
}

export enum Platform {
  NAVER = "naver",
  GOOGLE = "google",
}

export enum AgeRange {
  AGE_1418 = "1418",
  AGE_1924 = "1924",
  AGE_2529 = "2529",
  AGE_3034 = "3034",
  AGE_3539 = "3539",
  AGE_4049 = "4049",
  OVER_50 = "50+",
}

export enum UserGender {
  MALE = "M",
  FEMALE = "F",
}

export enum UserInterests {
  CROCHET = "crochet",
  KNITTING = "knitting",
}

export enum UserAbility {
  BEGINNER = "beginner",
  INTERMEDIATE = "intermediate",
  ADVANCED = "advanced",
  EXPERT = "expert",
}